pyproject.toml
[project]
name = "ref-audit-crossref"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.32.0",
  "python-dotenv>=1.0.1"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

